server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-telegram-bot==20.7
redis==5.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
structlog==24.1.0
python-dotenv==1.0.0
requests==2.31.0
